pyproject.toml
[tool.poetry]
name = "rag-website-chat"
version = "0.1.0"
description = ""
authors = ["Grey Lovelace"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.10"
pgvector = "^0.2.1"
langchain = {extras = ["llms"], version = "^0.0.267"}
instructorembedding = "^1.0.1"
python-dotenv = "^1.0.0"
psycopg2-binary = "^2.9.7"
unstructured = "^0.10.2"
tiktoken = "^0.4.0"
streamlit = "^1.25.0"
pypdf = "^3.15.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[[tool.poetry.source]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
default = false
secondary = true